+++ CHANGELOG.md
@@ -14,7 +14,10 @@
 - Update docs and configuration to use local npm scripts instead of global gulp commands
   - Example: `npm test` instead of `gulp test`
 - Update all npm packages
+- Add asset revisions to production builds using [gulp-rev](https://github.com/sindresorhus/gulp-rev)
+- Add favicon markup for layout templates
 - Remove ES5 JS option
+- Remove [gulp-notify](https://github.com/mikaelbr/gulp-notify) from build tasks
 
 ### v2.0.1: